Consolidate ListService test suites
Jira ticket: CAMS-526
